vst,resolution,buffer counter, buffer duration, resolution graph
This VideoJs Demo contains a VST that calculates the time in seconds between the click of play button untill the video begins to play. It contains a Video Resolution display so that the video tester knows what resolution they are watching on., this is for testing purposes. It also contains a Video Resolution chart that graphs the video resolution in pixels vs time , once again this is for testing purposes to see at what seconds in time it took for the video to reach the highest resolution. It also displays a buffered counter and a last buffered duration. This is once again for testing purposes.